requirements.txt
# # --- Core Web Framework ---
# fastapi[all]
# uvicorn[standard]
# python-multipart

# # --- Voice & Audio Processing ---
# openai-whisper
# faster-whisper           # optional, faster model inference
# sounddevice
# PyAudio
# scipy
# numpy
# pytz

# # --- AI & LangChain Ecosystem ---
# langchain-core
# langchain-astradb
# langchain-openai
# langdetect

# # --- Utilities & Parsing ---
# Jinja2                   # for templating if needed
# python-dotenv            # for env management
# webrtcvad

# --- Core Web Framework ---
fastapi==0.115.14
uvicorn==0.35.0
python-multipart==0.0.20
starlette==0.46.2

# --- Voice & Audio Processing ---
openai-whisper==20250625
faster-whisper==1.1.1
edge-tts==7.0.2
sounddevice==0.5.2

scipy
numpy
webrtcvad-wheels
ffmpeg
srt
av

# --- LangChain & AI Ecosystem ---
langchain-core==0.3.68
langchain==0.3.26
langchain-astradb==0.6.0
langchain-openai==0.3.27
langchain-community==0.3.27
langchain-text-splitters==0.3.8
langdetect==1.0.9
openai==1.93.0

huggingface-hub==0.33.2
tiktoken==0.9.0


# --- NLP / Parsing / Utilities ---
Jinja2==3.1.6
python-dotenv==1.1.1

jsonschema==4.24.0
pydantic==2.11.7


# --- Optional / Environment Tools ---

dataclasses-json==0.6.7

